+++ src/App.tsx
@@ -1,14 +1,14 @@
 import { createBrowserRouter, RouterProvider} from 'react-router-dom';
-import GrimList from './pages/GrimList';
-import WriteGrim from './pages/WriteGrim';
-import Main from './pages/Main';
-import AfterLogin from './pages/AfterLogin';
-import SignIn from './pages/SignIn';
-import SignUp from './pages/SignUp';
-import Manual from './pages/Manual';
-import DiarySearchList from './pages/DiarySearchList';
-import PrivatePages from './components/access/PrivatePages';
-import PublicPages from './components/access/PublicPages';
+import GrimList from './pages/DiaryList/GrimListPage';
+import WriteGrim from './pages/WriteDiary/WriteGrimPage';
+import Main from './pages/Main/NonMemberMainPage';
+import AfterLogin from './pages/Main/MemberMainPage';
+import SignIn from './pages/Auth/SignIn/SignInPage';
+import SignUp from './pages/Auth/SignUp/SignUpPage';
+import Manual from './pages/Manual/ManualPage';
+import DiarySearchList from './pages/Search/DiarySearchPage';
+import PrivatePages from './pages/Auth/components/PrivatePages';
+import PublicPages from './pages/Auth/components/PublicPages';
 import ErrorPage from './pages/ErrorPage';
 import { createTheme, ThemeProvider } from '@material-ui/core';
 import Root from './pages/Root';
@@ -26,18 +26,17 @@ const router= createBrowserRouter([
     element: <Root />,
     errorElement: <ErrorPage />,
     children: [
+      {path:'/', element:<PublicPages Component={Main} restricted/>},
       {path:'/signin', element:<PublicPages Component={SignIn} restricted/>},
       {path:'/signup', element:<PublicPages Component={SignUp} restricted/>},
-      // {path:'/write', element:<PrivatePages Component={WriteGrim}/>},
-      // {path:'/list', element:<PrivatePages Component={GrimList}/>},
+      {path:'/write', element:<PrivatePages Component={WriteGrim}/>},
       {path:'/write', element:<WriteGrim />},
       {
-        path:'/', element:<Navbar />,
+        path:'/', element:<PrivatePages Component={Navbar} />,
         children:[
-          {path:'/', element:<PublicPages Component={Main} restricted/>},
-          {path:'main', element:<AfterLogin />},
+          {path:'main', element:<PrivatePages Component={AfterLogin} />},
           {path:'about', element:<PublicPages Component={Manual} restricted />},
-          {path:'list', element:<GrimList />},
+          {path:'list', element:<PrivatePages Component={GrimList}/>},
           {path:'/search/:word', element:<DiarySearchList />}
         ]
       }
