+++ src/components/Loading.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styled from 'styled-components';
 function Loading(){
   return (
