+++ src/components/bookshape/BookShapeL.tsx
@@ -1,4 +1,4 @@
-import React, { PropsWithChildren } from 'react';
+import React from 'react';
 import styled from 'styled-components';
  
 type BookShape2LProps = {
